+++ requirements.txt
@@ -1,3 +1,4 @@
 fastapi[all]
 redis
-crfm-helm
+streamlit
+#crfm-helm
